+++ client/src/features/Products/components/useProductFilter.jsx
@@ -1,0 +1,10 @@
+import { useState } from 'react';
+
+const useProductFilter = () => {
+  const action = useProductActions()
+
+  return {
+  };
+};
+
+export default useProductFilter;
